+++ strongtalk/src/cpp/main/vm/compiler/Compiler.hpp
@@ -60,7 +60,7 @@ public:
     Node                *firstNode;                // the very first node of the intermediate representation
     PerformanceDebugger *reporter;                 // for reporting performance info
     StringOutputStream  *messages;                 // debug messages
-    std::array<int, static_cast<int>(InlineLimitType::LastLimit)> inlineLimit;                               // limits for current compilation
+    std::array<int, static_cast<std::size_t>(InlineLimitType::LastLimit)> inlineLimit;                               // limits for current compilation
 
 private:
     void initialize( RecompilationScope *remote_scope = nullptr );
